Removed enums from messages going across Rust FFI
The actual representation of enums is compiler dependent and unspecified.

+++ rust/src/spinner.rs
@@ -38,7 +38,7 @@ struct Channel {
     state: ChannelState,
     plan: [PlanLeg; constants::MAX_SPIN_PLAN_LEGS as usize],
     plan_len: usize,
-    elapsed_time_msecs: u64,
+    elapsed_time_msecs: u32,
     output_val_pct: f32,
 }
